pyproject.toml
[project]
name = "infinity-firestore-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin>=6.9.0",
    "google-cloud-firestore>=2.21.0",
    "pydantic>=2.11.7",
]
